requirements.txt
matplotlib==3.7.2
munkres==1.1.4
numpy==1.26.4
PyYAML==6.0
scikit_learn==1.4.1
scipy==1.10.1
torch==2.0.1
tqdm==4.65.0
